Going to try imagenet with all coefficients

+++ train_imagenet.py
@@ -186,7 +186,8 @@ if __name__ == "__main__":
         max_epochs=dict_args["epoch"],
         log_every_n_steps=10,
         gradient_clip_val=dict_args['clip_grad_norm'],
-        devices=-1
+        devices=-1,
+        precision="16-mixed"
     ) 
 
     # Trainer: train model
